Introduce =target-changed MI notification
So patch/RFC is reviving this 4 years old thread.

  http://thread.gmane.org/gmane.comp.gdb.patches/80427

I'll try to give an overview of the problem and the past discussion, so
you don't have to dig in the archives.

The train of thought goes like this:

The problem: In an MI front-end, the user can use the embedded GDB
console to type:

  (gdb) set $rbx = 0xabcd

Because of the lack of "register changed" notification, the registers
view in the frontend won't update.  The same happens if the user updates
a variable and that variable is stored in a register.  Both the
registers and variables view won't update.  When changing memory though,
we have the =memory-changed event, we see that something is missing.

Note that this only concerns cases where something (e.g. the user or a
script) changes the target state while it's stopped behind the
front-end's back.  Cases where the target is being run are not
considered here, since the front-end knows that it should refresh all of
its views, as their content could be outdated.

The solution proposed in the thread mentioned above was to add a
=register-changed notification (and that's what I would have
instinctively suggested as well).  However, in order for the frontend to
always be correctly sync'ed, we would need to handle some edge cases
which can be pretty hard to get right efficiently.  For example:

 - The user manually sets the value of a register of a frame other than
   the current frame.  It's possible for that register value to not be
   in a register at the moment, but to be saved on the stack, in which
   case GDB will update the value in the stack.  For the front-end to
   stay sync'ed, we would need to send a =register-changed event as well
   as a =memory-changed event, since both the registers and memory views
   need to be updated.  I guess that this one is not too hard, because
   if GDB changed the value, it knows whether it changed a register or a
   memory location.

 - The user manually changes a value in memory that happens to be a
   register from a previous frame that is saved on the stack.  This
   seems much harder to do.  From my very shy DWARF knowledge, I
   understand that it's easy to find out where a register belonging to a
   previous frame is stored (.debug_frames), but not the other way
   around.  In that case, the registers view would go out of sync.

Then, when you factor in memory-mapped registers, and other edge cases
we haven't imagined, this all sounds like a very complicated problem.

The other approach suggested in the thread was to not even try being
smart, and just emit a generic =target-changed notification when
anything was changed by the user behind the back of the MI frontend.
After all, changing anything (register or memory) can impact anything
else (registers, variables, frame stack, memory, etc.).  So all of those
views should be considered stale and be updated.

An argument for such an event was that front-ends need to do a full
refresh any time they single step or stop at a breakpoint.  These
operations happens in general much more frequently than the user
changing data by hand.  So doing a full refresh is probably not a big
deal in that case.

An in-between option that was considered was to add the
=register-changed event, but to make it clear in the doc that even
though this is the "origin" of the change, the front-end should refresh
pretty much everything.  I liked how Pedro illustrated it: "We may be
better off not giving ourselves and the frontends rope to hang ourselves
with.".  Indeed, what point is there to give some info to the front-end
but also tell it to not trust that info?

So that's why I think that a =target-changed event makes the most sense.
The =memory-changed event could co-exist with =target-changed for some
time, but if we go that route it would make sense to deprecate its usage.

What do you think?

gdb/ChangeLog:

	* mi/mi-cmd-var.c (mi_cmd_var_assign): Suppress target-changed
	MI notification.
	* mi/mi-interp.c (mi_target_changed): New declaration and
	function.
	(mi_interpreter_init): Attach to target_changed observer.
	* mi/mi-main.h (mi_suppress_notification) <target_changed>: New
	field.