Make mi-break.exp expect breakpoint commands output on the main UI
Even with the previous patch, we mi-break.exp still regressions when
tested with MI running on a secondary UI.

The problem is simply that the test set a breakpoint, and attaches
"print" commands to the breakpoint.  Since breakpoint commands always
run with the main UI as current UI, the breakpoint command's output
goes to the main UI.  So we need to tweak the test to expect it there.

gdb/ChangeLog:
yyyy-mm-dd  Pedro Alves  <palves@redhat.com>

	* gdb.mi/mi-break.exp (test_breakpoint_commands): Always expect
	breakpoint command's output on the main UI.
	* lib/mi-support.exp (gdb_main_spawn_id): Declare and comment.
	(mi_uncatched_gdb_exit): Clear gdb_main_spawn_id.
	(mi_gdb_start_separate_mi_tty, ): Set gdb_main_spawn_id.