arm fast tracepoints: Relocation of ARM instructions
This patch adds the possibility to relocate some common ARM-mode
instructions that may depend on the current value of the PC.

Branches are very frequent and are always pc-relative, so we make sure
to support pretty much all cases.  We calculate the new relative offset,
given the old and new instruction locations.

Other instructions are rejected as soon as they reference the PC.  We
judged these cases are not frequent enough to be worth the effort of
relocating.

gdb/gdbserver/ChangeLog:

	* linux-arm-low.c (struct arm_insn_reloc_data): Add orig_loc and
	new_loc fields.
	(arm_reloc_refs_pc_error): New function.
	(arm_reloc_alu_imm): Implement.
	(arm_reloc_alu_reg): Implement.
	(arm_reloc_alu_shifted_reg): Implement.
	(arm_reloc_b_bl_blx): Implement.
	(arm_reloc_block_xfer): Implement.
	(arm_reloc_bx_blx_reg): Implement.
	(arm_reloc_copro_load_store): Implement.
	(arm_reloc_extra_ld_st): Implement.
	(arm_reloc_ldr_str_ldrb_strb): Implement.
	(arm_reloc_preload): Implement.
	(arm_reloc_preload_reg): Implement.
	(arm_reloc_svc): Implement.
	(arm_reloc_undef): Implement.
	(arm_reloc_unpred): Implement.
	(copy_instruction_arm): Assign orig_loc and new_loc.

gdb/testsuite/ChangeLog:

	* gdb.trace/ftrace-arm-insn.S: New file.
	* gdb.trace/ftrace-arm-insn.c: New file.
	* gdb.trace/ftrace-arm-insn.exp: New file.