Change thread numbering in MI
Current gdb master has id="<global num>".  I do not see that changing
in the near/mid future, to keep compatibility with frontends.  If we
want to use the new format, it will have to be in a new field.  This
patch reverts the id field to be the globa number, and adds
the per-inferior-id field with the format
<inf number>.<not global thread number>.