mi: Restore original thread when specifying --thread or --thread-group
I am sending a first version of this patch, even though I am not
completely satisfied with it.  Hopefully I can get some input from the
community.

We are in the process of integrating Pedro's great new-ui/console work
[0] in Eclipse CDT.  With this, the hope is that GDB power users will
feel right at home with a console that behaves just like plain GDB in a
terminal.

Earlier, we found that if the current thread is X, passing --thread Y to
an MI command would leave Y as the current thread.  This is not a
problem for CDT itself, since it always uses --thread for MI commands.
However, it also affects the user selected thread.  So if the frontend
is doing some operations to refresh its views while the user types a
command, the thread can appear to change unexpectedly from the point of
view of the user.

So, to avoid changing the current selection under the user's feet while
they are typing a command,  --thread and --thread-group should leave the
inferior/thread selection in their original state.

This naive patch simply adds a restore_current_thread cleanup whenever
--thread or --thread-group is used.  The code that checks whether a
=thread-selected event should be emitted can be simplified.

It works okay for most commands, but I am worried about these corner
cases:

1. If the command is actually meant to change the current thread and
   --thread/--thread-group is specified, the command will have no
   effect.  Some front-ends might add --thread X for everything, so I
   think this is a plausible scenario:

   -thread-select --thread 3 4
   -interpreter-exec --thread 3 console "thread 4"

2. When a breakpoint is hit in all-stop, GDB's behavior is to switch to
   the thread that hit the breakpoint (which makes sense).  With a
   synchronous target, I have the feeling that it could inhibit this
   behavior, if the events go like this:

    1. The front-end issues "-exec-continue --thread-group i1".  Or with
       --thread 1, it doesn't matter because it's all-stop.
    2. We create a restorer_current_thread cleanup with the current thread
       (let's say #1)
    3. We enter the implementation of -exec-continue, we block somewhere
       deep in there because the target is sync
    4. A breakpoint is hit by thread #2, so the implementation of
       -exec-continue eventually returns.  It leaves thread #2 as the
       current thread, because it's the one that hit the breakpoint.
    5. We (wrongfully) restore thread #1 as being the current thread.

I can get around #1 by having an ugly global variable
"meant_to_change_current_thread" that can be set to mean that the thread
should not be restored to its original value, because we actually want
to change the user selected thread.  gdb_thread_select, for example,
would do that.  It really feels hackish though.

Perhaps #2 could be solved the same way, but I don't really know where
meant_to_change_current_thread would be set.  IOW, where does GDB take
the conscious decision to change/leave the user selected thread to the
thread that hit the breakpoint?

[0] https://sourceware.org/ml/gdb-patches/2016-05/msg00098.html
[1] https://www.sourceware.org/ml/gdb/2015-10/msg00073.html