mi: Restore original thread/frame when specifying --thread or --thread-group
We are in the process of integrating Pedro's great separate ui work [0]
in Eclipse CDT.  With this, the hope is that GDB users will feel right
at home with a console that behaves just like plain GDB in a terminal,
and appreciate the graphical support of CDT.

Earlier [1], we found that if the current thread is X, passing
"--thread Y" to an MI command would leave Y as the current thread.
This is not a problem for CDT itself, since it always uses --thread for
MI commands.  However, it also affects the user selected thread.  So the
internal front-end logic can unexpectedly change the currently selected
thread from the point of view of the user using the console.

So, to avoid changing the current selection under the user's feet while
they are typing a command,  --thread and --thread-group should leave the
inferior/thread/frame selection in their original state.

This patch adds a restore_current_thread cleanup whenever --thread or
--thread-group is used (using --frame implies using --thread).  To
support the case where the --thread flag would be used along with a
command that is meant to change the thread, it also adds a global
variable "command_changes_user_selected_ptid".  The GDB core can set
this variable to indicate that it wants to change the user-selected
thread.  When this variable is set, the cleanup is discarded so that the
newly selected thread is kept.  Because there is a single chain of
cleanups, I had to reorder things in mi_cmd_execute so that it was
possible to only discard this cleanup.

As a side effect, the code that checks whether a =thread-selected event
should be emitted was simplified.

This is intended to be a temporary solution, until we implement
something like what Pedro suggested in [2].

Regtested with --target_board=unix on Linux x86.

[0] https://sourceware.org/ml/gdb-patches/2016-05/msg00098.html
[1] https://www.sourceware.org/ml/gdb/2015-10/msg00073.html
[2] https://sourceware.org/ml/gdb-patches/2016-08/msg00031.html

gdb/ChangeLog:

	* infcmd.c (command_changes_user_selected_ptid): New variable.
	* inferior.h (command_changes_user_selected_ptid): New
	declaration.
	* infrun.c (init_wait_for_inferior): Set
	command_changes_user_selected_ptid.
	(normal_stop): Likewise.
	* mi/mi-main.c (mi_execute_command): Simplify computation of
	report_change.
	(mi_cmd_execute): Create restore current thread cleanup if
	one of --thread or --thread-group is used.  Move language and
	suppress_notification code blocks up.  Discard current thread
	cleanup if command_changes_user_selected_ptid is set.
	* thread.c (do_captured_thread_select): Set
	command_changes_user_selected_ptid.
	* top.c (prepare_execute_command): Unset
	command_changes_user_selected_ptid.