Add test for user context selection sync
New in v3:

	* No more sleeps in the C code, which makes the test a bit
	faster to run.
	* Use scheduler-locking to bring threads to their designated
	position in all-stop.
	* Consume MI output after "start" and many other places, so we
	don't deadlock or wrongfully match a previous MI event (thanks
	to Pedro for figuring this out).
	* Use [^\r\n]* instead of .* in many places to avoid matching
	too much.
	* The CLI event test of "-thread-select --thread 2 2" is
	commented out (to avoid having to wait the the timeout every
	time when we know it will never arrive) and kfailed.

This patch adds a test to verify that events are sent properly to all
UIs when the user selection context (inferior, thread, frame) changes.

The goal of the C test file is to provide two threads that are stopped with the
same predictable backtrace (so that we can test frame switching).  The barrier
helps us know when the child threads are started.  Then, scheduler-locking is
used to bring each thread one by one to the position we expect them to be
during the test.

gdb/testsuite/ChangeLog:

YYYY-MM-DD  Antoine Tremblay  <antoine.tremblay@ericsson.com>
YYYY-MM-DD  Simon Marchi  <simon.marchi@ericsson.com>

	PR gdb/20487
	* gdb.mi/user-selected-context-sync.exp: New file.
	* gdb.mi/user-selected-context-sync.c: New file.