Add "inferior id" and "id in inferior"  info to -thread-info output
New in v2:

  - Doc and NEWS
  - Rename fields, use "inf" instead of "tg"
  - Test output with a single inferior before adding a second

The CLI has been showing inferior-qualified thread ids for a while, but
front-ends have no way to do the same, as only the global thread ids are
included in the MI output.  In front-ends in general, but mostly in
those that include a gdb console, it would be interesting to show the
same thread id in the GUI than what "info threads" displays.  The user
would then see consistent thread ids across the two UIs.

This patch adds two extra fields to the -thread-info MI command output.
In this snippet of prettified -thread-info output, the lines marked with
a + are new:

    {
+     inf-id = "2",
+     id-in-inf = "1",
      id = "3",
      target-id = "Thread 0x2aaaaab04640 (LWP 3228)",
      name = "mi-qualified-th",
      ...
    },

Here, the thread has the global id 3 (monotonously increasing across all
inferiors) and the inferior-qualified id 2.1.  The global thread id is
still the one used throughout the rest of MI.  The new fields are simply
intended to be presented to the user.

We chose adding the new information as two separate fields rather than a
single field with the value "2.1".  We think that it gives the front-end
more flexibility about how it want to present the info.

I updated the doc to mention the new fields.  I changed the example to
something that shows usefulness of the new fields.  It also shows a
better variety of thread states.  Note that the old example was
erroneous, it did show a thread as "running" with a frame.

gdb/ChangeLog:

	* thread.c (print_thread_info_1): Rename field id-in-tg to
	qualified-id.  Add new fields id-in-inf and inf-id when printing
	on MI.
	* NEWS: Mention new fields.

gdb/testsuite/ChangeLog:

	* mi-qualified-thread-id.exp: New file.
	* mi-qualified-thread-id.c: New file.

gdb/doc/ChangeLog:

	* gdb.texinfo (GDB/MI Thread Information): Document fields
	inf-id and id-in-inf.  Update example.