Prevent turning record on while threads are running (PR 20869)
As stated in the bug description, trying to do "record" while threads
are running leads to a broken state.  I tried to see in the code if
there was anything to try to support the use case of enabling record
while the program is running, but didn't find anything.

When full/software recording is active, we are single stepping each
instruction.  Transitioning from the state where threads are running
freely (no recording) to recording enabled should therefore involve a
step where we stop them to initiate the single stepping.  I can't find
anything that looks like this, so my conclusion is that this was never a
supported use case (please correct me if I'm wrong).

The doc seems to confirm this:

  If the inferior is in the non-stop mode (see Non-Stop Mode) or in the
  asynchronous execution mode (see Background Execution), not all
  recording methods are available. The full recording method does not
  support these two modes.

The easy solution is to prevent the user for enabling record if a thread
is running.

No regression on the buildbot.

gdb/ChangeLog:

	* record-full.c (record_preopen): Error out if any thread is running.

gdb/testsuite:

	* gdb.reverse/record-while-running.c: New file.
	* gdb.reverse/record-while-running.exp: New file.