Make Python inferior-related internal functions return inferior_object*
This patch changes the functions that return an Inferior Python object
so they return an inferior_object pointer instead of a generic PyObject
pointer.  This should remove some casting back and forth from PyObject*
to inferior_object* in our codebase, helping a bit for type-safety.

On the other hand, this requires adding some casts from inferior_object*
to PyObject* when we deal with the Python API, which obviously accepts
only PyObject*.  I think this is a more appropriate place for casts
though.

It defines and uses gdbpy_inf_ref, a specialized version of
gdbpy_ref_base for inferior_object.

The next patch changes those functions again to make them return a
gdbpy_inf_ref, but I think that this intermediary step is useful to
verify the correctness, since the two changes have different goals
(type-safety for this one, reference counting for the next one).

One point I am not sure about is whether it breaks the gcc Python
checking plugin:

  extern PyTypeObject inferior_object_type
      CPYCHECKER_TYPE_OBJECT_FOR_TYPEDEF ("inferior_object");

It says _FOR_TYPEDEF, but the inferior_object type is no longer a
typedef, because I need to forward declare the type in
python-internal.h, and it's not possible to forward-declare a typedef.
I tried to use the cpychecker plugin but have to admit I am bit lost.  I
understand some manual fiddling with the gcc flags are required in order
load the plugin (which can be done with gcc-with-cpychecker).  It would
be nice (if it's not the case yet) to have a configure switch to
automatically enable that for the files in the python subdirectory.

gdb/ChangeLog:

	* python/py-ref.h (gdbpy_inf_ref): New typedef.
	* python/python-internal.h (struct inferior_object):
	Forward-declare.
	(thread_object) <inf_obj>: Change type to inferior_object*.
	(find_inferior_object): Change return type to inferior_object*.
	(inferior_to_inferior_object): Likewise.
	* python/py-exitedevent.c (create_exited_event_object): Use
	gdbpy_inf_ref.
	* python/py-inferior.c (inferior_object): Change from typedef +
	anonymous struct to simple struct.
	(inferior_to_inferior_object): Change return type to
	inferior_object*.
	(find_inferior_object): Likewise.
	(find_thread_object): Use gdbpy_inf_ref.
	(build_inferior_list): Likewise.
	(gdbpy_selected_inferior): Add cast.
	* python/py-infthread.c (thpy_get_inferior): Add cast.